Synoptic item named "ignoreRepaint" are considered to be unnamed and to use the extension ignoreRepaint=True
I dont check exactelly this behaviour in JDraw, but there is some
filters about that on the file fr.esrf.tangoatk.widget.jdraw.TangoSynopticHandler

Before this patch it was possible to select an item named "ignoreRepaint"
and which was a sub item of a real selectable item. Plus all this items
was selected at the same time all around the synoptic.