Fix synopic JDSwingObject background in some cases
Calling updateStyle when showQuality=True is irrevocable.
Using setShowQuality with False will not set back the object
with the right look and feel.

A call of updateStyle append sometime before the item
is fully initialized.

This patch disable the showQuality by default. It is set
to true later when we initialize the object (alarmEnabled JDraw
property).